eth_gateway/events/requirements/base.txt
faust
simple-settings
python-schema-registry-client
python-schema-registry-client[faust]
Faker
python-dotenv
